Cleaned up Sources

+++ Secure Pipes/SocketIO/SocketIO.m
@@ -872,7 +872,7 @@ NSString* const SocketIOException = @"SocketIOException";
             // stop connecting; no more delegate messages
             [connection cancel];
             
-            NSString *error = [NSString stringWithFormat:NSLocalizedString(@"Server returned status code %d", @""), statusCode];
+            NSString *error = [NSString stringWithFormat:NSLocalizedString(@"Server returned status code %ld", @""), (long)statusCode];
             NSDictionary *errorInfo = [NSDictionary dictionaryWithObject:error forKey:NSLocalizedDescriptionKey];
             NSError *statusError = [NSError errorWithDomain:SocketIOError
                                                        code:statusCode
